Title: 사칙연산 함수, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240111/사칙연산 함수/quadratic-operations-function.cpp
@@ -18,7 +18,7 @@ int main() {
         break;
     case '/':
         if (c != 0) {
-            std::cout << a << " / " << c << " = " << static_cast<double>(a) / c << std::endl;
+            std::cout << a << " / " << c << " = " << static_cast<int>(a) / c << std::endl;
         }
         break;
     }
